+++ json.h
@@ -81,8 +81,8 @@ typedef enum
 #define LIBJSON_DEFAULT_STACK_SIZE 256
 #define LIBJSON_DEFAULT_BUFFER_SIZE 4096
 
-typedef int (*json_parser_callback)(void *userdata, int type, const char *data, uint32_t length);
-typedef int (*json_printer_callback)(void *userdata, const char *s, uint32_t length);
+typedef int (*json_parser_callback)(void *userdata, int type, const char *data, size_t length);
+typedef int (*json_printer_callback)(void *userdata, const char *s, size_t length);
 
 typedef struct {
 	uint32_t buffer_initial_size;
@@ -171,27 +171,27 @@ int json_print_raw(json_printer *printer, int type, const char *data, uint32_t l
  * int, float, key, string need to be followed by a pointer to char and then a length.
  * if the length argument is -1, then the strlen function will use on the string argument.
  * the function call should always be terminated by -1 */
-int json_print_args(json_printer *, int (*f)(json_printer *, int, const char *, uint32_t), ...);
+int json_print_args(json_printer *, int (*f)(json_printer *, int, const char *, size_t), ...);
 
 /** callback from the parser_dom callback to create object and array */
-typedef void * (*json_parser_dom_begin_structure)(int, int, void *, int, const char *, int, void *);
+typedef void * (*json_parser_dom_begin_structure)(int, int, void *, int, const char *, size_t, void *);
 
 /** callback from the parser_dom callback to create object and array */
-typedef int (*json_parser_dom_end_structure)(int, int, const char *, int, void *, void *);
+typedef int (*json_parser_dom_end_structure)(int, int, const char *, size_t, void *, void *);
 
 /** callback from the parser_dom callback to create data values */
-typedef void * (*json_parser_dom_create_data)(int, const char *, uint32_t, void *);
+typedef void * (*json_parser_dom_create_data)(int, const char *, size_t, void *);
 
 /** callback from the parser helper callback to append a value to an object or array value
  * append(parent, key, key_length, val); */
-typedef int (*json_parser_dom_append)(void *, int, int, char *, uint32_t, void *, void *);
+typedef int (*json_parser_dom_append)(void *, int, int, char *, size_t, void *, void *);
 
 /** the json_parser_dom permits to create a DOM like tree easily through the
  * use of 3 callbacks where the user can choose the representation of the JSON values */
 typedef struct json_parser_dom
 {
 	/* object stack */
-	struct stack_elem { void *val; char *key; uint32_t key_length; int is_object_structure; int structure_value_count; } *stack;
+	struct stack_elem { void *val; char *key; size_t key_length; int is_object_structure; int structure_value_count; } *stack;
 	uint32_t stack_size;
 	uint32_t stack_offset;
 
@@ -221,7 +221,7 @@ int json_parser_dom_init(json_parser_dom *helper,
 int json_parser_dom_free(json_parser_dom *ctx);
 
 /** helper to parser callback that arrange parsing events into comprehensive JSON data structure */
-int json_parser_dom_callback(void *userdata, int type, const char *data, uint32_t length);
+int json_parser_dom_callback(void *userdata, int type, const char *data, size_t length);
 
 #ifdef __cplusplus
 }
